Allow the kubectl context user and cluster to be configurable
Currently the kubectl cluster and user are the same as passed into
k8s_deploy. It would be great to allow this to be explicitly
configurable and different than the cluster and user arguments. In
particular, this is handy for EKS with the awscli command

```shell
$ awseks update-kubeconfig --name foo
```

Where the context's cluster and user names == the arn of the EKS
cluster.